Add AutoMapper injection, and some exceptions

+++ Holtz_Academy.API/Services/BranchService.cs
@@ -1,5 +1,7 @@
 ﻿using Holtz_Academy.API.Data;
 using Holtz_Academy.API.Entities;
+using Holtz_Academy.API.Entities.ViewModels;
+using Holtz_Academy.API.Services.Exceptions;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
@@ -18,6 +20,7 @@ namespace Holtz_Academy.API.Services
         public async Task<List<Branch>> FindAllAsync()
         {
             return await _context.Branches.ToListAsync();
+            //return await _context.Branches.Select(x => new BranchViewModel(x.BranchReason, x.BranchStreet, x.BranchNeighborhood, x.BranchNumber , x.BranchCityName)).ToListAsync();
         }
         public async Task<Branch> FindByCodeAsync(int code)
         {
@@ -30,34 +33,42 @@ namespace Holtz_Academy.API.Services
                 _context.Branches.Add(branch);
                 await _context.SaveChangesAsync();
             }
-            catch
+            catch (DbUpdateException e) //comes from db
             {
-
+                throw new IntegrityException(e.Message);
             }
         }
         public async Task UpdateAsync(Branch branch)
         {
+            if (!await _context.Branches.AnyAsync(x => x.BranchCode == branch.BranchCode))
+            {
+                throw new NotFoundException("Don't have Branch with this code");
+            }
             try
             {
                 _context.Branches.Update(branch);
                 await _context.SaveChangesAsync();
             }
-            catch
+            catch (DbUpdateException e) //comes from db
             {
-
+                throw new IntegrityException(e.Message);
             }
         }
         public async Task RemoveAsync(int code)
         {
             Branch branch = await FindByCodeAsync(code);
+            if (branch == null)
+            {
+                throw new NotFoundException("Don't have Branch with this code");
+            }
             try
             {
                 _context.Branches.Remove(branch);
                 await _context.SaveChangesAsync();
             }
-            catch
+            catch (DbUpdateException e) //comes from db
             {
-
+                throw new IntegrityException(e.Message);
             }
         }
     }
