+++ Holtz_Academy.API/Services/EquipamentService.cs
@@ -1,5 +1,6 @@
 ﻿using Holtz_Academy.API.Data;
 using Holtz_Academy.API.Entities;
+using Holtz_Academy.API.Services.Exceptions;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
@@ -31,36 +32,44 @@ namespace Holtz_Academy.API.Services
                 _context.Equipaments.Add(equipament);
                 await _context.SaveChangesAsync();
             }
-            catch
+            catch (DbUpdateException e) //comes from db
             {
-
+                throw new IntegrityException(e.Message);
             }
         }
 
         public async Task UpdateAsync(Equipament equipament)
         {
+            if (!await _context.Equipaments.AnyAsync(x => x.EquipamentCode == equipament.EquipamentCode))
+            {
+                throw new NotFoundException("Don't have Equipament with this code");
+            }
             try
             {
                 _context.Equipaments.Update(equipament);
                 await _context.SaveChangesAsync();
             }
-            catch
+            catch (DbUpdateException e) //comes from db
             {
-
+                throw new IntegrityException(e.Message);
             }
         }
 
         public async Task RemoveAsync(int code)
         {
             Equipament equipament = await FindByCodeAsync(code);
+            if (equipament == null)
+            {
+                throw new NotFoundException("Don't have Equipament with this code");
+            }
             try
             {
                 _context.Equipaments.Remove(equipament);
                 await _context.SaveChangesAsync();
             }
-            catch
+            catch (DbUpdateException e) //comes from db
             {
-
+                throw new IntegrityException(e.Message);
             }
         }
     }
